+++ src/frontend/js/jwtSelector.js
