+++ src/frontend/js/main.js
@@ -9,9 +9,9 @@ import FastClick from 'fastclick';
 import React from 'react';
 import { render } from 'react-dom';
 
-import Root from './app/Root';
-import rootReducer from './app/reducers';
-import configureStore from './app/configureStore';
+import Root from './Root';
+import rootReducer from './reducers';
+import configureStore from './configureStore';
 
 const store = configureStore(rootReducer);
