+++ src/frontend/js/order/OrderDetails.js
@@ -8,7 +8,7 @@ import OrderItem from './OrderItem';
 import OrderProductItem from './OrderProductItem';
 import OrderStatusBadge from './OrderStatusBadge';
 import { getOrderById } from './reducer';
-import withFetchingOnMount from '../app/withFetchingOnMount';
+import withFetchingOnMount from '../../../common-client/fetch/withFetchingOnMount';
 import withWindowTitle from '../app/withWindowTitle';
 
 const OrderDetails = ({ order: { reference, date, total, status, products } }) => (
