+++ src/frontend/js/routes.js
@@ -1,7 +1,7 @@
 import App from './App';
-import productRoutes from '../product/routes';
-import orderRoutesFactory from '../order/routes';
-import userRoutes from '../user/routes';
+import productRoutes from './product/routes';
+import orderRoutesFactory from './order/routes';
+import userRoutes from './user/routes';
 
 export default store => ({
     childRoutes: [{
