+++ src/frontend/js/sagas.js
@@ -1,7 +1,7 @@
 import { fork } from 'redux-saga/effects';
-import userSagas from '../user/sagas';
-import productSagas from '../product/sagas';
-import orderSagas from '../order/sagas';
+import userSagas from './user/sagas';
+import productSagas from './product/sagas';
+import orderSagas from './order/sagas';
 
 export default function* () {
     yield fork(userSagas);
